开启人脸， 重置版本号 充值回调字段解析有问题待处理

+++ app/src/main/java/com/android/nanguo/app/api/old_data/WalletRechargeBean.java
@@ -4,7 +4,7 @@ public final class WalletRechargeBean {
     public String token;
     public String orderId;
     public String html;//":"若用户未签约，则返回签约页面，否则返回空字符串"
-    public double signStatus;//":"签约状态：0-未签约，1-已签约(int整型)"
+    public int signStatus;//":"签约状态：0-未签约，1-已签约(int整型)"
 
     public String getHtml() {
         return html;
@@ -14,11 +14,11 @@ public final class WalletRechargeBean {
         this.html = html;
     }
 
-    public double getSignStatus() {
+    public int getSignStatus() {
         return signStatus;
     }
 
-    public void setSignStatus(double signStatus) {
+    public void setSignStatus(int signStatus) {
         this.signStatus = signStatus;
     }
